move cnx code from grover to utils
