+++ qiskit_aqua/grover/__init__.py
@@ -17,4 +17,3 @@
 
 # for core development
 from .grover import Grover
-from .cnx import *
