+++ qiskit_aqua/utils/cnx.py
