+++ qiskit_aqua/utils/oracles/sat.py
@@ -18,12 +18,9 @@
 import itertools
 import operator
 import logging
-
 from qiskit import QuantumRegister, QuantumCircuit
-
 from qiskit_aqua.utils.oracles import Oracle
-import qiskit_aqua.grover.cnx
-
+from qiskit_aqua.utils.cnx import cnx
 
 logger = logging.getLogger(__name__)
